Fixed a tiny cast warning.
git-svn-id: https://s.snth.net/svn/rlk/thumb@2805 dcf3a027-e0e4-0310-89c9-ab9954db6d00

+++ proj/panoview/sph-cache.cpp
@@ -203,7 +203,7 @@ void sph_task::load_texture(TIFF *T, uint32 w, uint32 h, uint16 c, uint16 b)
         }
         else
         {
-            const uint32 S = TIFFScanlineSize(T);
+            const uint32 S = (uint32) TIFFScanlineSize(T);
 
             for (uint32 r = 0; r < h; ++r)
                 TIFFReadScanline(T, (uint8 *) p + r * S, r, 0);
